+++ tests/testthat/test_resample.R
@@ -451,6 +451,14 @@ test_that("multiple named measures", {
   expect_numeric(res[["classif.ce"]])
 })
 
+test_that("empty predictions", {
+  rr = resample(tsk("iris"), lrn("classif.debug", validate = 0.3, predict_sets = NULL), rsmp("holdout"))
+  preds = rr$predictions()
+  expect_equal(preds, list(list()))
+  pred = rr$prediction()
+  expect_equal(pred, list())
+})
+
 test_that("resample result works with not predicted predict set", {
   learner = lrn("classif.debug", predict_sets = "train")
   task = tsk("iris")