+++ NEWS.md
@@ -9,7 +9,10 @@
 * fix: column info is now checked for compatibility during `Learner$predict` (#943).
 * BREAKING CHANGE: the predict time of the learner now stores the cumulative duration for all predict sets (#992).
 * feat: `$internal_valid_task` can now be set to an `integer` vector.
-* refactor: Deprecated the `$divide()` method.
+* feat: Measures can now have an empty `$predict_sets` (#1094).
+  this is relevant for measures that only extract information from
+  the model of a learner (such as internal validation scores or AIC / BIC)
+* refactor: Deprecated the `$divide()` method
 * fix: `Task$cbind()` now works with non-standard primary keys  for `data.frames` (#961).
 * fix: Triggering of fallback learner now has log-level `"info"` instead of `"debug"` (#972).
 * feat: Added new measure `pinballs `.